Added Unlocking of Cards in Profile
- You can set the category which unlocks a card in the cms, throught the
<unlockedBy></unlockedBy> item. Possible values are: leaf, fruit, trunk
- Current threshold is set to 90% completed for a category
- Fixed some issues with comparisons and casting

+++ tree-app-android/app/src/main/java/de/lmu/treeapp/contentClasses/trees/Tree.java
@@ -85,6 +85,7 @@ public class Tree {
     }
 
     public float GetGameProgressionPercent(GameCategories category){
+
         switch (category){
             case leaf:
                 return GetGamesProgression(leafGamesIds, changeable.leafGamesCompleted) * 100;
@@ -110,13 +111,13 @@ public class Tree {
         int completed = 0;
         for (int i = 0; i < gamesCompleted.size(); i++){
             for (int j = 0; j < gamesTotal.size(); j++){
-                if (gamesTotal.get(j) == gamesCompleted.get(i)){
+                if (gamesTotal.get(j).intValue() == gamesCompleted.get(i).intValue()){
                     completed++;
                     break;
                 }
             }
         }
-        float value = completed/gamesTotal.size();
+        float value = (float)completed/(float)gamesTotal.size();
         return value;
     }
